+++ core-services/egov-user/src/main/java/org/egov/user/web/controller/LogoutController.java
@@ -3,14 +3,12 @@ package org.egov.user.web.controller;
 import org.egov.common.contract.response.Error;
 import org.egov.common.contract.response.ErrorResponse;
 import org.egov.common.contract.response.ResponseInfo;
+import org.egov.user.domain.model.TokenWrapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.common.OAuth2AccessToken;
 import org.springframework.security.oauth2.provider.token.TokenStore;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.Date;
 
@@ -31,7 +29,8 @@ public class LogoutController {
      * @throws Exception
      */
     @PostMapping("/_logout")
-    public ResponseInfo deleteToken(@RequestParam("access_token") String accessToken) throws Exception {
+    public ResponseInfo deleteToken(@RequestBody TokenWrapper tokenWrapper) throws Exception {
+        String accessToken = tokenWrapper.getAccessToken();
         OAuth2AccessToken redisToken = tokenStore.readAccessToken(accessToken);
         tokenStore.removeAccessToken(redisToken);
         return new ResponseInfo("", "", new Date().toString(), "", "", "Logout successfully");