+++ core-services/egov-workflow-v2/src/main/java/org/egov/wf/repository/rowmapper/WorkflowRowMapper.java
@@ -96,6 +96,7 @@ public class WorkflowRowMapper implements ResultSetExtractor<List<ProcessInstanc
                         .previousStatus(rs.getString("previousStatus"))
                         .moduleName(rs.getString("moduleName"))
                         .auditDetails(auditdetails)
+                        .rating(rs.getInt("rating"))
                         .build();
             }
             addChildrenToProperty(rs,processInstance);