+++ core-services/libraries/services-common/src/main/java/org/egov/common/contract/request/User.java
@@ -10,6 +10,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import javax.validation.constraints.Size;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -20,20 +22,27 @@ import lombok.ToString;
 public class User {
     private Long id;
 
+    @Size(max = 180)
     private String userName;
 
+    @Size(max = 250)
     private String name;
 
+    @Size(max = 50)
     private String type;
 
+    @Size(max = 150)
     private String mobileNumber;
 
+    @Size(max = 300)
     private String emailId;
 
     private List<Role> roles;
 
+    @Size(max = 256)
     private String tenantId;
-    
+
+    @Size(max = 36)
     private String uuid;
 }