+++ core-services/chatbot/CHANGELOG.md
@@ -3,6 +3,9 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.3 - 2021-02-26
+- Updated domain name in application.properties
+
 ## 1.1.2 - 2021-01-11
 - Updated PGR v2.0 API calls,request object and response object.