+++ core-services/egov-accesscontrol/CHANGELOG.md
@@ -3,6 +3,9 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.1 - 2021-02-26
+- Updated domain name in application.properties
+
 ## 1.1.0 - 2020-06-19
 - Added typescript definition generation plugin
 - Upgraded to `tracer:2.0.0-SNAPSHOT`