+++ core-services/egov-indexer/CHANGELOG.md
@@ -3,6 +3,10 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.3 - 2021-02-26
+
+- Updated domain name in application.properties
+
 ## 1.1.2
 
 - Added version handling and baselining.