+++ core-services/egov-pg-service/CHANGELOG.md
@@ -3,6 +3,9 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.2.1 - 2021-02-26
+- Updated domain name in application.properties
+
 ## 1.2.0 - 2020-10-20
 - Added support to make payment by other than the owner/citizen.