+++ core-services/egov-url-shortening/CHANGELOG.md
@@ -3,6 +3,9 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.0 - 2021-02-26
+- Updated domain name in application.properties
+
 ## 1.0.0 - 2020-06-12
 - Added typescript definition generation plugin
 - Upgraded to `tracer:2.0.0-SNAPSHOT`