+++ core-services/egov-user/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.2.3 - 2021-02-26
+- Updated domain name in application.properties
+- Added size validations
+
 ## 1.2.2 - 2020-01-12
 - Added field relationShip with guardian and refactoration of code.