+++ core-services/pdf-service/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.2 - 2021-02-26
+- Updated domain name in application.properties
+
 ## 1.1.1 -2021-01-12
 - Handled special character as input in PDF